refactor: movieList 컴포넌트 분리 - noResultItem을 컴포넌트로 분리하고 - for문으로 구성된 로직을 Array.from으로 대체하였음
